+++ src/coin-selection/lib/Convex/CoinSelection.hs
@@ -1,5 +1,7 @@
 {-# LANGUAGE BangPatterns       #-}
 {-# LANGUAGE DataKinds          #-}
+{-# LANGUAGE DeriveAnyClass     #-}
+{-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE GADTs              #-}
 {-# LANGUAGE LambdaCase         #-}
 {-# LANGUAGE NamedFieldPuns     #-}
@@ -12,6 +14,8 @@
 -}
 module Convex.CoinSelection(
   -- * Data types
+  CoinSelectionError(..),
+  bodyError,
   CSInputs(..),
   ERA,
   utxo,
@@ -52,6 +56,7 @@ import           Convex.Utxos          (BalanceChanges (..), UtxoSet (..))
 import qualified Convex.Utxos          as Utxos
 import           Convex.Wallet         (Wallet)
 import qualified Convex.Wallet         as Wallet
+import           Data.Aeson            (FromJSON, ToJSON)
 import           Data.Bifunctor        (Bifunctor (..))
 import           Data.Function         (on)
 import qualified Data.List             as List
@@ -62,6 +67,9 @@ import           Data.Maybe            (isNothing, listToMaybe, mapMaybe,
 import           Data.Ord              (Down (..))
 import           Data.Set              (Set)
 import qualified Data.Set              as Set
+import           Data.Text             (Text)
+import qualified Data.Text             as Text
+import           GHC.Generics          (Generic)
 
 type ERA = BabbageEra
 
@@ -94,17 +102,25 @@ makeLensesFor
 
 data CoinSelectionError =
   UnsupportedBalance (C.TxOutValue ERA)
-  | BodyError C.TxBodyError
+  | BodyError Text
   | NotEnoughAdaOnlyOutputsFor C.Lovelace
   | NotEnoughMixedOutputsFor{ valuesNeeded :: [(C.PolicyId, C.AssetName, C.Quantity)], valueProvided :: C.Value, txBalance :: C.Value }
-  deriving Show
+  deriving stock (Show, Generic)
+  deriving anyclass (ToJSON, FromJSON)
+
+bodyError :: C.TxBodyError -> CoinSelectionError
+bodyError = BodyError . Text.pack . C.displayError
 
 data BalancingError =
-  BalancingError C.TxBodyErrorAutoBalance
+  BalancingError Text
   | CheckMinUtxoValueError (C.TxOut C.CtxTx BabbageEra) C.Lovelace
   | BalanceCheckError BalancingError
   | ComputeBalanceChangeError
-  deriving Show
+  deriving stock (Show, Generic)
+  deriving anyclass (ToJSON, FromJSON)
+
+balancingError :: C.TxBodyErrorAutoBalance -> BalancingError
+balancingError = BalancingError . Text.pack . C.displayError
 
 {-| Perform transaction balancing
 -}
@@ -115,16 +131,16 @@ balanceTransactionBody systemStart eraHistory protocolParams stakePools CSInputs
       changeOutputLarge = mkChangeOutputFor ((2^(64 :: Integer)) - 1)
   -- append output instead of prepending
   txbody0 <-
-    first (BalancingError . C.TxBodyError) $ C.createAndValidateTransactionBody $ csiTxBody & appendTxOut changeOutputSmall
+    first (balancingError . C.TxBodyError) $ C.createAndValidateTransactionBody $ csiTxBody & appendTxOut changeOutputSmall
 
-  exUnitsMap <- first (BalancingError . C.TxBodyErrorValidityInterval) $
+  exUnitsMap <- first (balancingError . C.TxBodyErrorValidityInterval) $
                 C.evaluateTransactionExecutionUnits
                 systemStart (C.toLedgerEpochInfo eraHistory)
                 protocolParams
                 csiUtxo
                 txbody0
 
-  exUnitsMap' <- first BalancingError $
+  exUnitsMap' <- first balancingError $
     case Map.mapEither id exUnitsMap of
       (failures, exUnitsMap') ->
         handleExUnitsErrors C.ScriptValid failures exUnitsMap' -- TODO: should this take the script validity from csiTxBody?
@@ -132,7 +148,7 @@ balanceTransactionBody systemStart eraHistory protocolParams stakePools CSInputs
   let txbodycontent1 = substituteExecutionUnits exUnitsMap' csiTxBody
 
   -- append output instead of prepending
-  txbody1 <- first (BalancingError . C.TxBodyError)
+  txbody1 <- first (balancingError . C.TxBodyError)
               $ C.createAndValidateTransactionBody
               $ txbodycontent1
                   & set L.txFee (C.Lovelace (2^(32 :: Integer) - 1))
@@ -140,7 +156,7 @@ balanceTransactionBody systemStart eraHistory protocolParams stakePools CSInputs
 
   let !t_fee = C.evaluateTransactionFee protocolParams txbody1 csiNumWitnesses 0
 
-  txbody2 <- first (BalancingError . C.TxBodyError)
+  txbody2 <- first (balancingError . C.TxBodyError)
               $ C.createAndValidateTransactionBody
               $ txbodycontent1 & set L.txFee t_fee & appendTxOut csiChangeOutput
 
@@ -161,7 +177,7 @@ balanceTransactionBody systemStart eraHistory protocolParams stakePools CSInputs
     C.TxOutValue _ v -> do
       case C.valueToLovelace v of
         -- FIXME: Support non Ada assets
-        Nothing -> Left $ BalancingError $ C.TxBodyErrorNonAdaAssetsUnbalanced v
+        Nothing -> Left $ balancingError $ C.TxBodyErrorNonAdaAssetsUnbalanced v
         Just lvl  ->  do
           let op = csiChangeOutput & L._TxOut . _2 . L._TxOutValue . L._Value . at C.AdaAssetId <>~ (Just $ C.lovelaceToQuantity lvl)
           balanceCheck protocolParams op
@@ -172,7 +188,7 @@ balanceTransactionBody systemStart eraHistory protocolParams stakePools CSInputs
           & set L.txFee t_fee
           & over L.txOuts (accountForNoChange changeOutputBalance)
 
-  txbody3 <- first (BalancingError . C.TxBodyError) $ C.createAndValidateTransactionBody finalBodyContent
+  txbody3 <- first (balancingError . C.TxBodyError) $ C.createAndValidateTransactionBody finalBodyContent
 
   balances <- maybe (Left ComputeBalanceChangeError) Right (balanceChanges csiUtxo finalBodyContent)
 
@@ -211,7 +227,7 @@ balanceCheck pparams output =
     if view L._TxOutValue balance == mempty
       then return ()
       else do
-        when (C.txOutValueToLovelace balance < 0) (Left . BalancingError . C.TxBodyErrorAdaBalanceNegative $ C.txOutValueToLovelace balance)
+        when (C.txOutValueToLovelace balance < 0) (Left . balancingError . C.TxBodyErrorAdaBalanceNegative $ C.txOutValueToLovelace balance)
         bimap BalanceCheckError (const ()) $ checkMinUTxOValue output pparams
 
 updateRestWithChange :: C.TxOut C.CtxTx C.BabbageEra -> [C.TxOut C.CtxTx C.BabbageEra] -> [C.TxOut C.CtxTx C.BabbageEra]
@@ -414,7 +430,7 @@ runsScripts body =
 -}
 addMissingInputs :: Set PoolId -> C.BundledProtocolParameters BabbageEra -> C.UTxO ERA -> C.TxOut C.CtxTx C.BabbageEra -> UtxoSet ctx a -> TxBodyContent BuildTx ERA -> Either CoinSelectionError (TxBodyContent BuildTx ERA, C.TxOut C.CtxTx C.BabbageEra)
 addMissingInputs poolIds ledgerPPs utxo_ returnUTxO0 walletUtxo txBodyContent0 = do
-  txb <- first BodyError (C.createAndValidateTransactionBody txBodyContent0)
+  txb <- first bodyError (C.createAndValidateTransactionBody txBodyContent0)
   let bal = C.evaluateTransactionBalance ledgerPPs poolIds mempty utxo_ txb & view L._TxOutValue
       available = Utxos.removeUtxos (spentTxIns txBodyContent0) walletUtxo